refactor: ManualRecordService 제거 및 관련 로직을 ParticipantService로 통합
